Move the metrics into their own file
